+++ benchmark/load.php
@@ -12,11 +12,15 @@ use Formal\ORM\{
 use Innmind\OperatingSystem\Factory;
 use Innmind\Url\Url;
 use Fixtures\Formal\ORM\User;
+use Innmind\TimeContinuum\PointInTime;
 
 $os = Factory::build();
 $connection = $os->remote()->sql(Url::of("mysql://root:root@127.0.0.1:3306/example"));
 $aggregates = Aggregates::of(Types::of(
-    Type\PointInTimeType::of($os->clock()),
+    Type\Support::class(
+        PointInTime::class,
+        Type\PointInTimeType::new($os->clock()),
+    ),
 ));
 
 $manager = Manager::sql($connection, $aggregates);