+++ src/Definition/Type/PointInTimeType.php
@@ -26,14 +26,25 @@ use Innmind\Immutable\Maybe;
 final class PointInTimeType implements Type
 {
     private Clock $clock;
+    private Format $format;
 
-    private function __construct(Clock $clock)
+    private function __construct(Clock $clock, Format $format)
     {
         $this->clock = $clock;
+        $this->format = $format;
     }
 
     /**
      * @psalm-pure
+     */
+    public static function new(Clock $clock): self
+    {
+        return new self($clock, Format::new());
+    }
+
+    /**
+     * @psalm-pure
+     * @deprecated Use ::new() instead
      *
      * @return callable(Types, Concrete, ?Contains): Maybe<self>
      */
@@ -41,12 +52,12 @@ final class PointInTimeType implements Type
     {
         return static fn(Types $types, Concrete $type) => Maybe::just($type)
             ->filter(static fn($type) => $type->accepts(ClassName::of(PointInTime::class)))
-            ->map(static fn() => new self($clock));
+            ->map(static fn() => new self($clock, new Format));
     }
 
     public function normalize(mixed $value): null|string|int|bool
     {
-        return $value->format(new Format);
+        return $value->format($this->format);
     }
 
     public function denormalize(null|string|int|bool $value): mixed
@@ -57,7 +68,7 @@ final class PointInTimeType implements Type
 
         return $this
             ->clock
-            ->at($value, new Format)
+            ->at($value, $this->format)
             ->match(
                 static fn($point) => $point,
                 static fn() => throw new \LogicException("'$value' is not a date"),